removed some warnings due to UInt--Size ;; some documentation-updates
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@4304 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/MATH/STATISTICS/Histogram.h
@@ -295,7 +295,7 @@ namespace OpenMS
 				}
 				if (val == max_)
 				{
-					return (bins_.size()-1);
+					return UInt(bins_.size()-1);
 				}
 				else
 				{
